requirements.txt
asgiref==3.7.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.5
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
phonenumbers==8.13.21
Pillow==10.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.3.7